requirements.txt
flask==3.0.2
flask-login==0.6.3
mysql-connector-python==8.3.0
python-dotenv==1.0.1
bcrypt==4.1.2
flask-sqlalchemy==3.1.1
pymysql==1.1.0
flask-migrate==4.0.7
flask-script==2.0.6 